MAINT: dedicated namespaces for APIs
