+++ imageio/v3.py
@@ -1,0 +1,6 @@
+from .core.v3_api import imopen
+from .core.v3_api import imread
+from .core.v3_api import imwrite
+from .core.v3_api import imiter
+
+__all__ = ["imopen", "imread", "imwrite", "imiter"]
